Add Health Data example (#162)
This change adds an example of a TypeChat translator with history. Specifically, the translator keeps the previous agent responses and adds them the prompt, allowing the model to take them into account. In the health data example, the agent helps the user fill out information on their health history, current conditions and medication. The agent asks clarifying questions and keeps track of the data provided in the course of a conversation. This example is ported from the TypeChat.Net implementation.

Note: This sample works best with GPT 4